+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- ShowAvailability property to show stock instead of lowStock or highStock message
+
 ## [0.1.0] - 2021-03-08
 ### Added
 - CSS Handles custom classes